сделал код ревью
