+++ scr/services/FilmGetService.php
@@ -1,35 +1,24 @@
 <?php
 namespace WhatToWatch\services;
+use WhatToWatch\services\FilmRepositoryInterface;
 
-use GuzzleHttp;
-
-interface FilmRepository {
-    public function save($id);
-    public function getInfo();
-    public function showInfo();
-}
-
-class FilmGetService implements FilmRepository
+class FilmGetService
 {
-    private $id;
-    private $info;
-    private $apiKey = '6defab64';
-    private $site = 'https://www.omdbapi.com';
+    private $repository;
 
-    public function save($id)
+    public function __construct(FilmRepositoryInterface $repository)
     {
-        $this->id = $id;
+        $this->repository = $repository;
+    }
+
+    public function save(string $filmId)
+    {
+        return $this->repository->save($filmId);
     }
 
     public function getInfo()
     {
-        $client = new \GuzzleHttp\Client();
-        $response = $client->get($this->site . '/?apikey=' . $this->apiKey . '&i=' . $this->id);
-        $this->info = $response->getBody();
+        return $this->repository->getInfo();
     }
 
-    public function showInfo()
-    {
-        return $this->info;
-    }
 }
