+++ scr/services/FilmRepositoryInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace WhatToWatch\services;
+interface FilmRepositoryInterface
+{
+    public function save($id);
+    public function getInfo();
+}
